conan v2 support
