--- recipes/sophus/all/test_cmakedeps/CMakeLists.txt
@@ -1,8 +1,0 @@
-cmake_minimum_required(VERSION 3.15)
-project(test_package)
-
-find_package(Sophus REQUIRED CONFIG)
-
-add_executable(${PROJECT_NAME} test_package.cpp)
-target_compile_features(${PROJECT_NAME} PRIVATE cxx_std_11)
-target_link_libraries(${PROJECT_NAME} Sophus::Sophus)
